finish feat ext

+++ openood/utils/logger.py
@@ -1,9 +1,7 @@
 import errno
-import json
 import os
 import os.path as osp
 import sys
-from re import T
 
 import yaml
 
@@ -73,33 +71,38 @@ def setup_logger(config):
     Args:
         config: all configs of the experiment
     """
-    output = osp.join(config.output_dir, config.exp_name)
-    if not config.force_merge and osp.isdir(output):
-        ans = input('Exp dir already exists, merge it? (y/n)')
-        if ans in ['yes', 'Yes', 'YES', 'y', 'Y', 'can']:
-            pass
-        elif ans in ['no', 'No', 'NO', 'n', 'N']:
-            print('Quitting the process...', flush=True)
-            quit()
-        else:
-            raise ValueError('Unexpected Input.')
-    else:
-        print('Output directory path: {}'.format(output), flush=True)
-        os.makedirs(output, exist_ok=True)
-
-    # Save config
-    # FIXME: saved config file is not beautified.
     print('------------------ Config --------------------------', flush=True)
     print(config, flush=True)
     print('----------------------------------------------------', flush=True)
-    config_save_path = osp.join(output, 'config.yml')
-    with open(config_save_path, 'w') as f:
-        yaml.dump(config,
-                  f,
-                  default_flow_style=False,
-                  sort_keys=False,
-                  indent=2)
+    output = osp.join(config.output_dir, config.exp_name)
 
-    # save log file
-    fpath = osp.join(output, 'log.txt')
-    sys.stdout = Logger(fpath)
+    if config.save_output:
+
+        if not config.force_merge and osp.isdir(output):
+            ans = input('Exp dir already exists, merge it? (y/n)')
+            if ans in ['yes', 'Yes', 'YES', 'y', 'Y', 'can']:
+                pass
+            elif ans in ['no', 'No', 'NO', 'n', 'N']:
+                print('Quitting the process...', flush=True)
+                quit()
+            else:
+                raise ValueError('Unexpected Input.')
+
+        else:
+            print('Output directory path: {}'.format(output), flush=True)
+            os.makedirs(output, exist_ok=True)
+            # Save config
+            # FIXME: saved config file is not beautified.
+            config_save_path = osp.join(output, 'config.yml')
+            with open(config_save_path, 'w') as f:
+                yaml.dump(config,
+                          f,
+                          default_flow_style=False,
+                          sort_keys=False,
+                          indent=2)
+            # save log file
+            fpath = osp.join(output, 'log.txt')
+            sys.stdout = Logger(fpath)
+
+    else:
+        print('No output directory.', flush=True)
